8213255121804376606
ManifestID: 8213255121804376606
Date: 2026-06-29T00:18:59+08:00
Version: OC-254439

+++ Assembly-CSharp/NavType.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-public enum NavType : byte
+public enum NavType : short
 {
 	Floor,
 	LeftWall,
